fss: carry FSS altitude at full width so only the clamp narrows it (todo/55)
The FSS wire altitude (fss_message_asset_command::getAltitude, uint32_t feet)
was narrowed to uint16_t in FSSCommandTarget before the [floor, cap] clamp ever
saw it. A malformed or oversized value could wrap during that cast to a small
altitude the safety clamp would then accept as legitimate -- e.g. 65536 ft
wrapping to 0, clamped up to the floor, so a "far too high" command became "as
low as allowed".

Widen the path (FSSCommandTarget::altitude, IMAV/MAV::setAltitude,
mav_connection::commandAltitude, clamp_command_altitude) to uint32_t so
clamp_command_altitude is the single narrowing authority: an oversized value
converts to a large metre count and pins to the cap, never wraps.

Tests: an oversized altitude reaches setAltitude intact through the state
machine, and clamp_command_altitude pins 65536/65569/UINT32_MAX to the cap.

+++ src/mav/imav.hpp
@@ -34,7 +34,7 @@ class IMAV
     virtual auto disarm () -> bool = 0;
     virtual auto terminate () -> bool = 0;
     virtual void gotoPosition (Point to) = 0;
-    virtual auto setAltitude (uint16_t alt) -> bool = 0;
+    virtual auto setAltitude (uint32_t alt) -> bool = 0;
     virtual auto getCurrentPosition () -> Point = 0;
     virtual void registerMavCommsStatusCB (notify_mav_comms_cb cb) = 0;
 };
